chore: migrated hoare-law-z3-solver type aliases to separate ts models file for clarity
